Avoid importing from kolibri.utils.conf
Side-effects from importing this code are affecting performance and
causing unexpected behaviour with creating KOLIBRI_HOME from a template.

Ref #19
