+++ build-aux/flatpak/python3-kolibri.json
@@ -4,7 +4,8 @@
     "build-commands": [
         "pip3 install --exists-action=i --no-index --find-links=\"file://${PWD}\" --prefix=${FLATPAK_DEST} kolibri",
         "patch /app/lib/python3.7/site-packages/kolibri/utils/server.py server.py.diff",
-        "patch /app/lib/python3.7/site-packages/kolibri/utils/cli.py cli.py.diff"
+        "patch /app/lib/python3.7/site-packages/kolibri/utils/cli.py cli.py.diff",
+        "patch /app/lib/python3.7/site-packages/kolibri/utils/logger.py logger.py.diff"
     ],
     "sources": [
         {
@@ -19,6 +20,10 @@
         {
             "type": "file",
             "path": "cli.py.diff"
+        },
+        {
+            "type": "file",
+            "path": "logger.py.diff"
         }
     ]
 }
